+++ praktikum1/src/main/java/aufgabe1/utils/NetworkEdge.java
@@ -6,22 +6,22 @@ import org.jgrapht.graph.DefaultWeightedEdge;
  * Created by schlegel11 on 07.12.14.
  */
 public class NetworkEdge extends DefaultWeightedEdge {
-    private double capacity;
-    private double flow;
+    private int capacity;
+    private int flow;
 
-    public double getCapacity() {
+    public int getCapacity() {
         return capacity;
     }
 
-    public void setCapacity(double capacity) {
+    public void setCapacity(int capacity) {
         this.capacity = capacity;
     }
 
-    public double getFlow() {
+    public int getFlow() {
         return flow;
     }
 
-    public void setFlow(double flow) {
+    public void setFlow(int flow) {
         this.flow = flow;
     }
 }
